refactor: 유저 리포지토리 추상화
