+++ src/modules/base/essay/core/essay.service.ts
@@ -10,7 +10,7 @@ import { ToolService } from '../../../utils/tool/tool.service';
 import { AwsService } from '../../../adapters/aws/core/aws.service';
 import { ReviewService } from '../../../extensions/management/review/review.service';
 import { StoryService } from '../../../extensions/essay/story/story.service';
-import { UserService } from '../../user/user.service';
+import { UserService } from '../../user/core/user.service';
 import { TagService } from '../../../extensions/essay/tag/tag.service';
 import { FollowService } from '../../../extensions/user/follow/follow.service';
 import { BadgeService } from '../../../extensions/essay/badge/core/badge.service';
