+++ src/modules/base/essay/tests/essay.service.spec.ts
@@ -5,7 +5,7 @@ import { ToolService } from '../../../utils/tool/tool.service';
 import { AwsService } from '../../../adapters/aws/core/aws.service';
 import { ReviewService } from '../../../extensions/management/review/review.service';
 import { StoryService } from '../../../extensions/essay/story/story.service';
-import { UserService } from '../../user/user.service';
+import { UserService } from '../../user/core/user.service';
 import { TagService } from '../../../extensions/essay/tag/tag.service';
 import { FollowService } from '../../../extensions/user/follow/follow.service';
 import { BadgeService } from '../../../extensions/essay/badge/core/badge.service';
@@ -29,7 +29,7 @@ jest.mock('../../util/util.service');
 jest.mock('../../aws/core/aws.service');
 jest.mock('../../review/review.service');
 jest.mock('../../story/story.service');
-jest.mock('../../user/user.service');
+jest.mock('../../user/core/user.service');
 jest.mock('../../tag/tag.service');
 jest.mock('../../follow/follow.service');
 jest.mock('../../badge/core/badge.service');
