+++ src/modules/base/user/user.module.ts
@@ -7,16 +7,17 @@ import { MailModule } from '../../utils/mail/mail.module';
 import { AwsModule } from '../../adapters/aws/aws.module';
 import { EssayModule } from '../essay/essay.module';
 import { ToolModule } from '../../utils/tool/tool.module';
-import { UserController } from './user.controller';
-import { UserRepository } from './user.repository';
-import { UserService } from './user.service';
+import { UserController } from './api/user.controller';
+import { UserRepository } from './infrastructure/user.repository';
+import { UserService } from './core/user.service';
 import { User } from '../../../entities/user.entity';
 import { Essay } from '../../../entities/essay.entity';
 import { NicknameModule } from '../../utils/nickname/nickname.module';
 import { DeactivationReason } from '../../../entities/deactivationReason.entity';
 import { BullModule } from '@nestjs/bull';
 import { ConfigService } from '@nestjs/config';
-import { UserProcessor } from './user.processor';
+import { UserProcessor } from './core/user.processor';
+import { IUserRepository } from './infrastructure/iuser.repository';
 
 @Module({
   imports: [
@@ -41,7 +42,12 @@ import { UserProcessor } from './user.processor';
     forwardRef(() => EssayModule),
   ],
   controllers: [UserController],
-  providers: [UserService, UserRepository, UserProcessor, strategies.JwtStrategy],
-  exports: [UserService, UserRepository],
+  providers: [
+    UserService,
+    { provide: 'IUserRepository', useClass: UserRepository },
+    UserProcessor,
+    strategies.JwtStrategy,
+  ],
+  exports: [UserService, { provide: 'IUserRepository', useClass: UserRepository }],
 })
 export class UserModule {}
